Refactor visit_GreaterThanNode Function More
Refactor code for the visit_GreaterThanNode function in the interpreter by moving (check_x, float) and (check_y, float) to the same line as (check_x, int) and (check_y, int) with an or. Also, merger the if and elif statements together with an or. This reduced lines of code.