Refactor visit_GreaterThanEqualNode Function Again
Refactor code for the visit_GreaterThanEqualNode function in the interpreter by moving the isinstance check for ((check_x, int) and (check_y, float)) or ((check_x, float) and (check_y, int)) to the same line as isinstance for ((check_x, int) and (check_y, int)) or ((check_x, float) and (check_y, float)) with an or. Also, merger the if and elif statements together with an or. This reduced lines of code.