Refactor visit_NotEqualNode Function Again
Refactor code for the visit_NotEqualNode function in the interpreter by moving the isinstance check for ((check_x, int) and (check_y, float)) or ((check_x, float) and (check_y, int)) to the same line as isinstance for ((check_x, int) and (check_y, int)) or ((check_x, float) and (check_y, float)) with an or. Also, merger the other if and elif statements together. This reduced lines of code.