Refactor Code in visit_TypeNotEqualNode Function
Refactor code for the visit_TypeNotEqualNode function in the interpreter by changing elif check_x != check_y: to else: because that line was just the opposite of a previous line in the code. Also remove line with  if (check_x == check_y) or (check_x != check_y): because the isinstance line above it is already checking for everything we need for part of the function.