Move KEYWORDS Token to ERROR_WORDS Token
Move KEYWORDS token to ERROR_WORDS token so that McFly could output an error message for non keywords. In addition, add the ERROR_WORDS token check to each starting letter that is already used so that they can also output that error message for non keywords starting with that letter.